method getusernamesuggestion
