+++ imports/services/index.js
@@ -1,6 +1,7 @@
 // import config from './moleculer.config';
 import Queue from './services/queue.js';
 import Authorization from './services/authorization';
+import User from './services/user';
 import { ServiceBroker } from 'moleculer';
 import { Meteor } from 'meteor/meteor';
 import { RocketChat } from 'meteor/rocketchat:lib';
@@ -13,6 +14,7 @@ const broker = new ServiceBroker({
 });
 broker.createService(Queue);
 broker.createService(Authorization);
+broker.createService(User);
 
 RocketChat.Services = broker;
